requirements.txt
webdriver-manager==4.0.2
selenium==4.29.0
python-dotenv==1.0.1
discord.py==2.5.2
